# apps/indigo-api-gateway/application/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API gateway for Indigo Server.",
        "title": "Indigo Server",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/responses": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new LLM response for the given input. Supports multiple input types including text, images, files, web search, and more.\n\n**Supported Input Types:**\n- `text`: Plain text input\n- `image`: Image input (URL or base64)\n- `file`: File input by file ID\n- `web_search`: Web search input\n- `file_search`: File search input\n- `streaming`: Streaming input\n- `function_calls`: Function calls input\n- `reasoning`: Reasoning input\n\n**Example Request:**\n```json\n{\n\"model\": \"gpt-4\",\n\"input\": {\n\"type\": \"text\",\n\"text\": \"Hello, how are you?\"\n},\n\"max_tokens\": 100,\n\"temperature\": 0.7,\n\"stream\": false,\n\"background\": false\n}\n```\n\n**Response Format:**\nThe response uses embedded structure where all fields are at the top level:\n- `jan_status`: Jan API status code (optional)\n- `id`: Response identifier\n- `object`: Object type (\"response\")\n- `created`: Unix timestamp\n- `model`: Model used\n- `status`: Response status\n- `input`: Input data\n- `output`: Generated output\n\n**Example Response:**\n```json\n{\n\"jan_status\": \"000000\",\n\"id\": \"resp_1234567890\",\n\"object\": \"response\",\n\"created\": 1234567890,\n\"model\": \"gpt-4\",\n\"status\": \"completed\",\n\"input\": {\n\"type\": \"text\",\n\"text\": \"Hello, how are you?\"\n},\n\"output\": {\n\"type\": \"text\",\n\"text\": {\n\"value\": \"I'm doing well, thank you!\"\n}\n}\n}\n```\n\n**Response Status:**\n- `completed`: Response generation finished successfully\n- `processing`: Response is being generated\n- `failed`: Response generation failed\n- `cancelled`: Response was cancelled",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Responses API"
                ],
                "summary": "Create a response",
                "parameters": [
                    {
                        "description": "Request payload containing model, input, and generation parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_requests.CreateResponseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created response",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.Response"
                        }
                    },
                    "202": {
                        "description": "Response accepted for background processing",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/responses/{response_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves an LLM response by its ID. Returns the complete response object with embedded structure where all fields are at the top level.\n\n**Response Format:**\nThe response uses embedded structure where all fields are at the top level:\n- `jan_status`: Jan API status code (optional)\n- `id`: Response identifier\n- `object`: Object type (\"response\")\n- `created`: Unix timestamp\n- `model`: Model used\n- `status`: Response status\n- `input`: Input data\n- `output`: Generated output",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Responses API"
                ],
                "summary": "Get a response",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unique identifier of the response",
                        "name": "response_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response details",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Response not found",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes an LLM response by its ID. Returns the deleted response object with embedded structure where all fields are at the top level.\n\n**Response Format:**\nThe response uses embedded structure where all fields are at the top level:\n- `jan_status`: Jan API status code (optional)\n- `id`: Response identifier\n- `object`: Object type (\"response\")\n- `created`: Unix timestamp\n- `model`: Model used\n- `status`: Response status (will be \"cancelled\")\n- `input`: Input data\n- `cancelled_at`: Cancellation timestamp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Responses API"
                ],
                "summary": "Delete a response",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unique identifier of the response",
                        "name": "response_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted response",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Response not found",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "menlo_ai_indigo-api-gateway_app_interfaces_http_responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}